+++ server/core/logger.js
@@ -1,5 +1,6 @@
 const winston = require('winston');
 const _ = require('lodash');
+const moment = require('moment');
 
 let customLogLevels = _.clone(winston.config.syslog.levels);
 delete customLogLevels.emerg;
@@ -9,10 +10,12 @@ customLogColors.report = 'cyan';
 winston.addColors(customLogColors);
 let logger = new (winston.Logger)({
 	levels: customLogLevels,
-  exitOnError: false,
 	transports: [
 		new (winston.transports.Console)({
-			colorize: 'all'
+			colorize: 'all',
+      timestamp: function () {
+			  return moment().format('YYYY-MM-DD HH:mm:ss');
+      },
 		}),
   ]
 });