+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this module are documented in this file.
 poloLender application adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.12.0] - Unreleased
+
+### Added
+- timestamp for console log output
+- display the minimum offer amount in the "status" tab 
+
 ## [0.11.0] - 2017-06-26
 
 ### Added